Use `e2e` and `test` to distinguish test types
