+++ .vscode-test.mjs
@@ -1,7 +1,7 @@
 // https://github.com/microsoft/vscode-test-cli
 import { defineConfig } from '@vscode/test-cli';
 export default defineConfig({
-    files: ['out/src/**/*.test.js'],
+    files: ['out/src/**/*.e2e.js'],
     // https://mochajs.org/#command-line-usage
     // https://github.com/mochajs/mocha/tree/main/example/config
     mocha: {
