+++ src/parser/parser.e2e.ts
@@ -2,7 +2,7 @@ import { getDocument } from '../test/utils';
 import * as assert from 'assert';
 import * as path from 'path';
 import * as vscode from 'vscode';
-import { Parser } from '../parser/parser';
+import { Parser } from './parser';
 
 suite('Parser', () => {
     suite('detectVariableByLine', () => {
